# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Base Plate API",
        "description": "A product of Appetiser",
        "contact": {
            "name": "API Support",
            "email": "dennis.agulo@appetiser.com.au"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://128.199.193.26/api/v1",
            "description": "Base Endpoint"
        }
    ],
    "paths": {
...
        "/user/verify": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Verify user's email address",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ApiUsersController::verifyEmail",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email Address",
                                        "type": "string"
                                    },
                                    "verification_code": {
                                        "description": "Verification Code",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Invalid Input"
                    },
                    "400": {
                        "description": "Invalid Token"
                    },
                    "401": {
                        "description": "Token Expired"
                    },
                    "404": {
                        "description": "User Not Found / Token Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "200": {
                        "description": "Request OK"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/user/resend/email": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Resend verification email with token",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ApiUsersController::resendEmail",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Email Address",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Invalid Input"
                    },
                    "400": {
                        "description": "Invalid Token / Bad Request"
                    },
                    "401": {
                        "description": "Token Expired"
                    },
                    "404": {
                        "description": "User Not Found / Token Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "200": {
                        "description": "Request OK"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/user/update": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update user's information",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ApiUsersController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "description": "First Name",
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "description": "Last Name",
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "country": {
                                        "description": "Country",
                                        "type": "string",
                                        "example": "Philippines"
                                    },
                                    "mobile_number": {
                                        "description": "Mobile Number",
                                        "type": "string",
                                        "example": "+63 9277468888"
                                    },
                                    "address": {
                                        "description": "Address",
                                        "type": "string",
                                        "example": "216 Cabreros Street Cebu City"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid Token"
                    },
                    "401": {
                        "description": "Token Expired"
                    },
                    "404": {
                        "description": "User Not Found / Token Not Found"
                    },
                    "422": {
                        "description": "Invalid Input"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "200": {
                        "description": "Request OK"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/user/upload": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Upload profile photo",
                "operationId": "App\\Http\\Controllers\\API\\V1\\ApiUsersController::uploadProfilePhoto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "photo": {
                                        "description": "Image File",
                                        "type": "string",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid Token"
                    },
                    "401": {
                        "description": "Token Expired"
                    },
                    "404": {
                        "description": "Token Not Found"
                    },
                    "422": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "200": {
                        "description": "Request OK"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}